chore: add openapi tag to satisfy global sdk (#670)
