+++ pkg/client/.speakeasy/gen.lock
@@ -1,12 +1,12 @@
 lockVersion: 2.0.0
 id: a9ac79e1-e429-4ee3-96c4-ec973f19bec3
 management:
-  docChecksum: a75fbcee4705d8d6603062086b542987
+  docChecksum: 66467c4916574601e600551d938bede8
   docVersion: v1
   speakeasyVersion: 1.351.0
   generationVersion: 2.384.1
-  releaseVersion: 0.5.1
-  configChecksum: c054a60832022bfc610a27df523d0e92
+  releaseVersion: 0.5.2
+  configChecksum: cbb85382da20101bd1a7a4eb33420e1d
 features:
   go:
     additionalDependencies: 0.1.0
@@ -53,8 +53,6 @@ generatedFiles:
   - internal/utils/retries.go
   - internal/utils/security.go
   - internal/utils/utils.go
-  - /models/operations/v2getinfo.go
-  - /models/operations/getmetrics.go
   - /models/operations/getinfo.go
   - /models/operations/getledgerinfo.go
   - /models/operations/countaccounts.go
@@ -75,6 +73,8 @@ generatedFiles:
   - /models/operations/getbalances.go
   - /models/operations/getbalancesaggregated.go
   - /models/operations/listlogs.go
+  - /models/operations/v2getinfo.go
+  - /models/operations/getmetrics.go
   - /models/operations/v2listledgers.go
   - /models/operations/v2getledger.go
   - /models/operations/v2createledger.go
@@ -100,15 +100,11 @@ generatedFiles:
   - /models/operations/v2listlogs.go
   - /models/operations/v2importlogs.go
   - /models/operations/v2exportlogs.go
-  - /models/sdkerrors/v2errorresponse.go
-  - /models/sdkerrors/errorresponse.go
-  - /models/components/v2errorsenum.go
-  - /models/components/v2configinforesponse.go
-  - /models/components/httpmetadata.go
   - /models/components/configinforesponse.go
   - /models/components/configinfo.go
   - /models/components/config.go
   - /models/components/ledgerstorage.go
+  - /models/components/httpmetadata.go
   - /models/components/errorsenum.go
   - /models/components/ledgerinforesponse.go
   - /models/components/ledgerinfo.go
@@ -137,6 +133,8 @@ generatedFiles:
   - /models/components/aggregatebalancesresponse.go
   - /models/components/logscursorresponse.go
   - /models/components/log.go
+  - /models/components/v2errorsenum.go
+  - /models/components/v2configinforesponse.go
   - /models/components/v2ledgerlistresponse.go
   - /models/components/v2ledger.go
   - /models/components/v2getledgerresponse.go
@@ -172,8 +170,8 @@ generatedFiles:
   - /models/components/v2logscursorresponse.go
   - /models/components/v2log.go
   - /models/components/security.go
-  - docs/models/operations/v2getinforesponse.md
-  - docs/models/operations/getmetricsresponse.md
+  - /models/sdkerrors/errorresponse.go
+  - /models/sdkerrors/v2errorresponse.go
   - docs/models/operations/getinforesponse.md
   - docs/models/operations/getledgerinforequest.md
   - docs/models/operations/getledgerinforesponse.md
@@ -214,6 +212,8 @@ generatedFiles:
   - docs/models/operations/getbalancesaggregatedresponse.md
   - docs/models/operations/listlogsrequest.md
   - docs/models/operations/listlogsresponse.md
+  - docs/models/operations/v2getinforesponse.md
+  - docs/models/operations/getmetricsresponse.md
   - docs/models/operations/v2listledgersrequest.md
   - docs/models/operations/v2listledgersresponse.md
   - docs/models/operations/v2getledgerrequest.md
@@ -265,15 +265,11 @@ generatedFiles:
   - docs/models/operations/v2importlogsresponse.md
   - docs/models/operations/v2exportlogsrequest.md
   - docs/models/operations/v2exportlogsresponse.md
-  - docs/models/sdkerrors/v2errorresponse.md
-  - docs/models/sdkerrors/errorresponse.md
-  - docs/models/components/v2errorsenum.md
-  - docs/models/components/v2configinforesponse.md
-  - docs/models/components/httpmetadata.md
   - docs/models/components/configinforesponse.md
   - docs/models/components/configinfo.md
   - docs/models/components/config.md
   - docs/models/components/ledgerstorage.md
+  - docs/models/components/httpmetadata.md
   - docs/models/components/errorsenum.md
   - docs/models/components/ledgerinforesponse.md
   - docs/models/components/storage.md
@@ -311,6 +307,8 @@ generatedFiles:
   - docs/models/components/logscursorresponse.md
   - docs/models/components/type.md
   - docs/models/components/log.md
+  - docs/models/components/v2errorsenum.md
+  - docs/models/components/v2configinforesponse.md
   - docs/models/components/v2ledgerlistresponsecursor.md
   - docs/models/components/v2ledgerlistresponse.md
   - docs/models/components/v2ledger.md
@@ -363,9 +361,11 @@ generatedFiles:
   - docs/models/components/v2logtype.md
   - docs/models/components/v2log.md
   - docs/models/components/security.md
+  - docs/models/sdkerrors/errorresponse.md
+  - docs/models/sdkerrors/v2errorresponse.md
   - docs/sdks/formance/README.md
-  - docs/models/operations/option.md
   - docs/sdks/ledger/README.md
+  - docs/models/operations/option.md
   - docs/sdks/v1/README.md
   - docs/sdks/v2/README.md
   - USAGE.md
