+++ pkg/client/README.md
@@ -69,11 +69,11 @@ func main() {
 	)
 
 	ctx := context.Background()
-	res, err := s.Ledger.GetInfo(ctx)
+	res, err := s.Ledger.V1.GetInfo(ctx)
 	if err != nil {
 		log.Fatal(err)
 	}
-	if res.V2ConfigInfoResponse != nil {
+	if res.ConfigInfoResponse != nil {
 		// handle response
 	}
 }
@@ -83,11 +83,6 @@ func main() {
 
 <!-- Start Available Resources and Operations [operations] -->
 ## Available Resources and Operations
-
-### [Ledger](docs/sdks/ledger/README.md)
-
-* [GetInfo](docs/sdks/ledger/README.md#getinfo) - Show server information
-* [GetMetrics](docs/sdks/ledger/README.md#getmetrics) - Read in memory metrics
 
 ### [Ledger.V1](docs/sdks/v1/README.md)
 
@@ -114,6 +109,8 @@ func main() {
 
 ### [Ledger.V2](docs/sdks/v2/README.md)
 
+* [GetInfo](docs/sdks/v2/README.md#getinfo) - Show server information
+* [GetMetrics](docs/sdks/v2/README.md#getmetrics) - Read in memory metrics
 * [ListLedgers](docs/sdks/v2/README.md#listledgers) - List ledgers
 * [GetLedger](docs/sdks/v2/README.md#getledger) - Get a ledger
 * [CreateLedger](docs/sdks/v2/README.md#createledger) - Create a ledger
@@ -168,7 +165,7 @@ func main() {
 	)
 
 	ctx := context.Background()
-	res, err := s.Ledger.GetInfo(ctx, operations.WithRetries(
+	res, err := s.Ledger.V1.GetInfo(ctx, operations.WithRetries(
 		retry.Config{
 			Strategy: "backoff",
 			Backoff: &retry.BackoffStrategy{
@@ -182,7 +179,7 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-	if res.V2ConfigInfoResponse != nil {
+	if res.ConfigInfoResponse != nil {
 		// handle response
 	}
 }
@@ -221,11 +218,11 @@ func main() {
 	)
 
 	ctx := context.Background()
-	res, err := s.Ledger.GetInfo(ctx)
+	res, err := s.Ledger.V1.GetInfo(ctx)
 	if err != nil {
 		log.Fatal(err)
 	}
-	if res.V2ConfigInfoResponse != nil {
+	if res.ConfigInfoResponse != nil {
 		// handle response
 	}
 }
@@ -238,10 +235,10 @@ func main() {
 
 Handling errors in this SDK should largely match your expectations.  All operations return a response object or an error, they will never return both.  When specified by the OpenAPI spec document, the SDK will return the appropriate subclass.
 
-| Error Object              | Status Code               | Content Type              |
-| ------------------------- | ------------------------- | ------------------------- |
-| sdkerrors.V2ErrorResponse | default                   | application/json          |
-| sdkerrors.SDKError        | 4xx-5xx                   | */*                       |
+| Error Object            | Status Code             | Content Type            |
+| ----------------------- | ----------------------- | ----------------------- |
+| sdkerrors.ErrorResponse | default                 | application/json        |
+| sdkerrors.SDKError      | 4xx-5xx                 | */*                     |
 
 ### Example
 
@@ -266,10 +263,10 @@ func main() {
 	)
 
 	ctx := context.Background()
-	res, err := s.Ledger.GetInfo(ctx)
+	res, err := s.Ledger.V1.GetInfo(ctx)
 	if err != nil {
 
-		var e *sdkerrors.V2ErrorResponse
+		var e *sdkerrors.ErrorResponse
 		if errors.As(err, &e) {
 			// handle error
 			log.Fatal(e.Error())
@@ -319,11 +316,11 @@ func main() {
 	)
 
 	ctx := context.Background()
-	res, err := s.Ledger.GetInfo(ctx)
+	res, err := s.Ledger.V1.GetInfo(ctx)
 	if err != nil {
 		log.Fatal(err)
 	}
-	if res.V2ConfigInfoResponse != nil {
+	if res.ConfigInfoResponse != nil {
 		// handle response
 	}
 }
@@ -354,11 +351,11 @@ func main() {
 	)
 
 	ctx := context.Background()
-	res, err := s.Ledger.GetInfo(ctx)
+	res, err := s.Ledger.V1.GetInfo(ctx)
 	if err != nil {
 		log.Fatal(err)
 	}
-	if res.V2ConfigInfoResponse != nil {
+	if res.ConfigInfoResponse != nil {
 		// handle response
 	}
 }
@@ -433,11 +430,11 @@ func main() {
 	)
 
 	ctx := context.Background()
-	res, err := s.Ledger.GetInfo(ctx)
+	res, err := s.Ledger.V1.GetInfo(ctx)
 	if err != nil {
 		log.Fatal(err)
 	}
-	if res.V2ConfigInfoResponse != nil {
+	if res.ConfigInfoResponse != nil {
 		// handle response
 	}
 }
