+++ pkg/client/docs/sdks/v2/README.md
@@ -3,6 +3,8 @@
 
 ### Available Operations
 
+* [GetInfo](#getinfo) - Show server information
+* [GetMetrics](#getmetrics) - Read in memory metrics
 * [ListLedgers](#listledgers) - List ledgers
 * [GetLedger](#getledger) - Get a ledger
 * [CreateLedger](#createledger) - Create a ledger
@@ -28,6 +30,108 @@
 * [ListLogs](#listlogs) - List the logs from a ledger
 * [ImportLogs](#importlogs)
 * [ExportLogs](#exportlogs) - Export logs
+
+## GetInfo
+
+Show server information
+
+### Example Usage
+
+```go
+package main
+
+import(
+	"github.com/formancehq/ledger/pkg/client/models/components"
+	"github.com/formancehq/ledger/pkg/client"
+	"context"
+	"log"
+)
+
+func main() {
+    s := client.New(
+        client.WithSecurity(components.Security{
+            ClientID: "",
+            ClientSecret: "",
+        }),
+    )
+
+    ctx := context.Background()
+    res, err := s.Ledger.V2.GetInfo(ctx)
+    if err != nil {
+        log.Fatal(err)
+    }
+    if res.V2ConfigInfoResponse != nil {
+        // handle response
+    }
+}
+```
+
+### Parameters
+
+| Parameter                                                | Type                                                     | Required                                                 | Description                                              |
+| -------------------------------------------------------- | -------------------------------------------------------- | -------------------------------------------------------- | -------------------------------------------------------- |
+| `ctx`                                                    | [context.Context](https://pkg.go.dev/context#Context)    | :heavy_check_mark:                                       | The context to use for the request.                      |
+| `opts`                                                   | [][operations.Option](../../models/operations/option.md) | :heavy_minus_sign:                                       | The options for this request.                            |
+
+
+### Response
+
+**[*operations.V2GetInfoResponse](../../models/operations/v2getinforesponse.md), error**
+| Error Object              | Status Code               | Content Type              |
+| ------------------------- | ------------------------- | ------------------------- |
+| sdkerrors.V2ErrorResponse | default                   | application/json          |
+| sdkerrors.SDKError        | 4xx-5xx                   | */*                       |
+
+## GetMetrics
+
+Read in memory metrics
+
+### Example Usage
+
+```go
+package main
+
+import(
+	"github.com/formancehq/ledger/pkg/client/models/components"
+	"github.com/formancehq/ledger/pkg/client"
+	"context"
+	"log"
+)
+
+func main() {
+    s := client.New(
+        client.WithSecurity(components.Security{
+            ClientID: "",
+            ClientSecret: "",
+        }),
+    )
+
+    ctx := context.Background()
+    res, err := s.Ledger.V2.GetMetrics(ctx)
+    if err != nil {
+        log.Fatal(err)
+    }
+    if res.Object != nil {
+        // handle response
+    }
+}
+```
+
+### Parameters
+
+| Parameter                                                | Type                                                     | Required                                                 | Description                                              |
+| -------------------------------------------------------- | -------------------------------------------------------- | -------------------------------------------------------- | -------------------------------------------------------- |
+| `ctx`                                                    | [context.Context](https://pkg.go.dev/context#Context)    | :heavy_check_mark:                                       | The context to use for the request.                      |
+| `opts`                                                   | [][operations.Option](../../models/operations/option.md) | :heavy_minus_sign:                                       | The options for this request.                            |
+
+
+### Response
+
+**[*operations.GetMetricsResponse](../../models/operations/getmetricsresponse.md), error**
+| Error Object              | Status Code               | Content Type              |
+| ------------------------- | ------------------------- | ------------------------- |
+| sdkerrors.V2ErrorResponse | default                   | application/json          |
+| sdkerrors.SDKError        | 4xx-5xx                   | */*                       |
 
 ## ListLedgers
